Actually use file provider
