+++ py/core/providers/database/base.py
@@ -1,9 +1,6 @@
-from contextlib import contextmanager
-from typing import Any, Callable, Iterator, Optional, Union
+from typing import Any, Optional, Union
 
-from psycopg2.extensions import connection as pg_connection
-from psycopg2.extensions import lobject
-from sqlalchemy import TextClause, text
+from sqlalchemy import text
 
 from .vecs import Client
 
@@ -78,72 +75,13 @@ class QueryBuilder:
 
 
 class DatabaseMixin:
-    vx: Client
-
     def _get_table_name(self, base_name: str) -> str:
-        raise NotImplementedError("Subclasses must implement this method")
-
-    @contextmanager
-    def get_session(self) -> Iterator[Any]:
         raise NotImplementedError("Subclasses must implement this method")
 
     def execute_query(
         self, query: Union[str, text], params: Optional[dict[str, Any]] = None
-    ) -> Any:
-        with self.get_session() as sess:
-            if isinstance(query, str):
-                query = text(query)
-            result = sess.execute(query, params or {})
-            sess.commit()
-            return result
-
-    @contextmanager
-    def get_connection(self) -> Iterator[pg_connection]:
-        with self.get_session() as sess:
-            connection = sess.connection().connection
-            try:
-                yield connection
-            finally:
-                connection.close()
-
-    def execute_with_connection(
-        self, operation: Callable[[pg_connection], Any]
-    ) -> Any:
-        with self.get_connection() as conn:
-            return operation(conn)
+    ):
+        raise NotImplementedError("Subclasses must implement this method")
 
     def create_table(self):
         raise NotImplementedError("Subclasses must implement this method")
-
-    @contextmanager
-    def get_lobject(self, oid: int = 0, mode: str = "rb") -> Iterator[lobject]:
-        with self.get_connection() as conn:
-            lobj = conn.lobject(oid, mode)
-            try:
-                yield lobj
-            finally:
-                lobj.close()
-
-    def read_lob(self, oid: int) -> bytes:
-        with self.get_lobject(oid, "rb") as lobj:
-            return lobj.read()
-
-    def delete_lob(self, oid: int) -> None:
-        with self.get_connection() as conn:
-            conn.lobject(oid, "wb").unlink()
-
-    def execute_lob_query(
-        self,
-        query: Union[str, TextClause],
-        params: Optional[dict[str, Any]] = None,
-    ) -> Any:
-        with self.get_connection() as conn:
-            with conn.cursor() as cur:
-                if isinstance(query, TextClause):
-                    query = query.text
-                elif not isinstance(query, str):
-                    raise TypeError(
-                        "Query must be a string or SQLAlchemy TextClause object"
-                    )
-                cur.execute(query, params or {})
-                return cur.fetchone()
