+++ py/core/providers/file/__init__.py
@@ -1,0 +1,3 @@
+from .postgres import PostgresFileProvider
+
+__all__ = ["PostgresFileProvider"]
