+++ py/core/providers/file/postgres.py
@@ -1,53 +1,100 @@
 import io
 import logging
 from contextlib import contextmanager
-from typing import BinaryIO, Optional
+from typing import Any, BinaryIO, Callable, Iterator, Optional, Union
 from uuid import UUID
 
 from psycopg2 import Error as PostgresError
+from psycopg2.extensions import connection as pg_connection
+from psycopg2.extensions import lobject
 from sqlalchemy import text
+from sqlalchemy.sql.elements import TextClause
 
 from core.base import R2RException
-from core.base.providers import DatabaseConfig, FileProvider
-from core.providers.database.vecs import Client
-
-from .base import DatabaseMixin
+from core.base.providers import FileProvider
+from core.providers.database.postgres import PostgresDBProvider
 
 logger = logging.getLogger(__name__)
 
 
-class FileInfo:
-    def __init__(
+class PostgresFileProvider(FileProvider):
+    def __init__(self, db_provider: PostgresDBProvider):
+        super().__init__()
+        self.db_provider = db_provider
+        self.vx = db_provider.vx
+        self.create_table()
+
+    def _get_table_name(self, base_name: str) -> str:
+        return f"{base_name}"
+
+    @contextmanager
+    def get_session(self) -> Iterator[Any]:
+        with self.vx.Session() as sess:
+            yield sess
+
+    def execute_query(
+        self, query: Union[str, text], params: Optional[dict[str, Any]] = None
+    ) -> Any:
+        with self.get_session() as sess:
+            if isinstance(query, str):
+                query = text(query)
+            result = sess.execute(query, params or {})
+            sess.commit()
+            return result
+
+    @contextmanager
+    def get_connection(self) -> Iterator[pg_connection]:
+        with self.get_session() as sess:
+            connection = sess.connection().connection
+            try:
+                yield connection
+            finally:
+                connection.close()
+
+    def execute_with_connection(
+        self, operation: Callable[[pg_connection], Any]
+    ) -> Any:
+        with self.get_connection() as conn:
+            return operation(conn)
+
+    @contextmanager
+    def get_lobject(self, oid: int = 0, mode: str = "rb") -> Iterator[lobject]:
+        with self.get_connection() as conn:
+            lobj = conn.lobject(oid, mode)
+            try:
+                yield lobj
+            finally:
+                lobj.close()
+
+    def read_lob(self, oid: int) -> bytes:
+        with self.get_lobject(oid, "rb") as lobj:
+            return lobj.read()
+
+    def delete_lob(self, oid: int) -> None:
+        with self.get_connection() as conn:
+            conn.lobject(oid, "wb").unlink()
+
+    def execute_lob_query(
         self,
-        document_id: UUID,
-        file_name: str,
-        file_oid: int,
-        file_size: int,
-        file_type: Optional[str],
-        created_at: str,
-        updated_at: str,
-    ):
-        self.document_id = document_id
-        self.file_name = file_name
-        self.file_oid = file_oid
-        self.file_size = file_size
-        self.file_type = file_type
-        self.created_at = created_at
-        self.updated_at = updated_at
+        query: Union[str, TextClause],
+        params: Optional[dict[str, Any]] = None,
+    ) -> Any:
+        with self.get_connection() as conn:
+            with conn.cursor() as cur:
+                if isinstance(query, TextClause):
+                    query = query.text
+                elif not isinstance(query, str):
+                    raise TypeError(
+                        "Query must be a string or SQLAlchemy TextClause object"
+                    )
+                cur.execute(query, params or {})
+                return cur.fetchone()
 
-    def convert_to_db_entry(self):
-        return {
-            "document_id": self.document_id,
-            "file_name": self.file_name,
-            "file_oid": self.file_oid,
-            "file_size": self.file_size,
-            "file_type": self.file_type,
-            "created_at": self.created_at,
-            "updated_at": self.updated_at,
-        }
+    @contextmanager
+    def get_session(self):
+        with self.vx.Session() as sess:
+            yield sess
 
-
-class FileMixin(DatabaseMixin):
     def create_table(self):
         query = text(
             f"""
@@ -60,8 +107,6 @@ class FileMixin(DatabaseMixin):
             created_at TIMESTAMPTZ DEFAULT NOW(),
             updated_at TIMESTAMPTZ DEFAULT NOW()
         );
-        CREATE INDEX IF NOT EXISTS idx_file_name_{self.collection_name}
-        ON {self._get_table_name('file_storage')} (file_name);
         """
         )
         try:
@@ -110,7 +155,6 @@ class FileMixin(DatabaseMixin):
                 message=f"Failed to upsert file for document {document_id}",
             )
 
-    # TODO: Move this to the DatabaseMixin class. Causing problems due to async context management with lobjects
     def store_file(self, document_id, file_name, file_content, file_type=None):
         file_size = file_content.getbuffer().nbytes
 
@@ -137,7 +181,7 @@ class FileMixin(DatabaseMixin):
             raise R2RException(
                 status_code=500,
                 message=f"Failed to store file for document {document_id}: {e}",
-            )
+            ) from e
 
     def retrieve_file(
         self, document_id: UUID
@@ -240,45 +284,3 @@ class FileMixin(DatabaseMixin):
             }
             for row in results
         ]
-
-
-class PostgresFileProvider(FileProvider, FileMixin):
-    def __init__(
-        self, config: DatabaseConfig, vx: Client, collection_name: str
-    ):
-        FileProvider.__init__(self, config)
-        self.vx = vx
-        self.collection_name = collection_name
-
-    @contextmanager
-    def get_session(self):
-        with self.vx.Session() as sess:
-            yield sess
-
-    def create_table(self):
-        return FileMixin.create_table(self)
-
-    def store_file(self, document_id, file_name, file_content, file_type=None):
-        return FileMixin.store_file(
-            self, document_id, file_name, file_content, file_type
-        )
-
-    def retrieve_file(self, document_id):
-        return FileMixin.retrieve_file(self, document_id)
-
-    def delete_file(self, document_id):
-        return self.delete_file(document_id)
-
-    def get_files_overview(
-        self,
-        filter_document_ids=None,
-        filter_file_names=None,
-        offset=0,
-        limit=100,
-    ):
-        return FileMixin.get_files_overview(
-            self, filter_document_ids, filter_file_names, offset, limit
-        )
-
-    def _get_table_name(self, base_name: str) -> str:
-        return f"{base_name}_{self.collection_name}"
